pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bi-dashboard-validator"
version = "0.1.0"
description = "AI-assisted dashboard validation system for BI platforms."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "jinja2>=3.1.4",
  "openai>=1.58.0",
  "platformdirs>=4.3.6",
  "playwright>=1.50.0",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "redis>=5.2.1",
  "rq>=1.16.2",
  "rq-scheduler>=0.13.1",
  "sqlalchemy>=2.0.36",
  "structlog>=24.4.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pyinstaller>=6.11.1",
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
]

[project.scripts]
bi-validator = "bi_validator.launcher:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
